requirements.txt
pandas>=1.3.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=1.0.0
